Utility function for circuit connection check is added.
Initial version only checks for unconnected input and output pins of the nodes in argument. Unconnected output pins are reported as warnings since only a partition of some nodes' outputs might be used for some circuits. Unconnected input pins are reported as errors and they interrupt execution since a node with an undefined input pin is not realizable in any hardware.